item classification

+++ app/Http/Controllers/ItemController.php
@@ -11,6 +11,9 @@ use Redirect;
 use DB;
 use App\ItemGroup;
 use App\JxItem;
+use Paginator;
+use App\Service;
+use App\SabayItemTypes;
 class ItemController extends Controller
 {
     /**
@@ -18,10 +21,28 @@ class ItemController extends Controller
      *
      * @return Response
      */
+    public function getgametype($gametype){
+        $gameservice = Service::lists('code','code');
+        if(strcasecmp($gametype,'ak')==0){
+             $item = Item::paginate(env('PAGINATION'))->setPath('item');
+       }
+       elseif(strcasecmp($gametype,'jx2')==0){ 
+        $item = DB::connection('odbc')
+        ->select(DB::raw("select *,item.id as id,item.name as name,type.name as typename,groups.name as groupname from sabay_items as item INNER JOIN sabay_item_types as type on item.type_id=type.id INNER JOIN sabay_item_groups as groups on item.group_id = groups.id "));
+       }
+       else{
+        return Redirect::back();
+       }
+       //return Redirect::route('item');
+        return view('item.index',['items'=>$item,'gametype'=>$gametype,'gameservice'=>$gameservice]);
+    }
     public function index()
     {
+        //$jxitem = DB::connection('odbc')->select(DB::raw("select * from item"));
+        $gameservice = Service::lists('code','code');
         $item = Item::paginate(env('PAGINATION'))->setPath('item');
-        return view('item.index',['items'=>$item]);
+       
+        return view('item.index',['items'=>$item,'gametype'=>'ak','gameservice'=>$gameservice]);
     }
 
     /**
@@ -31,23 +52,10 @@ class ItemController extends Controller
      */
     public function create()
     {
-        /*$id = 1;
-        $results = DB::connection('odbc')->select(DB::raw("select * from item02 where itemguid=".$id));
-        if(!$results){
-            return 'jk';
-        }
-        else{
-            
-            foreach($results as $r){
-                return $r->itemguid;
-            }
-        }
-        //$r = DB::connection('odbc')->table('item02')->get(); 
-        dd($results);
-      // $jxitems = JxItem::All();
-        //dd($jxitems);*/
+        $gameservice = Service::lists('code','code');
+        $itemtype = SabayItemTypes::lists('name','id');
         $itemgroup = ItemGroup::lists('name','id');
-        return view('item.newitem',['itemgroup'=>$itemgroup]);
+        return view('item.newitem',['itemgroup'=>$itemgroup,'itemtype'=>$itemtype,'gameservice'=>$gameservice]);
     }
 
     /**
@@ -59,6 +67,7 @@ class ItemController extends Controller
     public function store(Request $request)
     {
         //strcasecmp($time,"period") ==0
+
         if(strcasecmp($request->gametypes,'ak') == 0 ){
                 $item =  new Item;
                 $id = Item::where('id',$request->id)->first();
@@ -66,14 +75,15 @@ class ItemController extends Controller
                     return Redirect::back()->withErrors('ID Already Exist');
                 }
                 else{
-                     $item->id = $request->id;
+                    //return 'hi';
+                        $item->id = $request->id;
                         $item->name = $request->name;
-                        $item->type = $request->itemtype;
+                        $item->type_id = $request->itemtype;
+                        $item->group_id = $request->itemgroup;
                         $item->duration = $request->duration;
                         $item->price= $request->price;
-                         if(is_null($request->dateadded)){$item->date_added = "";}
-                            else{$item->date_added = $request->dateadded;}
-                        $item->group_id = $request->itemgroup;
+                        if(is_null($request->dateadded)){$item->date_added = "";}
+                        else{$item->date_added = $request->dateadded;}
                         $item->save();
         }
     }
@@ -82,12 +92,11 @@ class ItemController extends Controller
         $date_added = strtotime($request->dateadded);
         $date_added = date('Y-m-d',$date_added);
 
-        $checkid = DB::connection('odbc')->select(DB::raw("select * from item where ID=".$request->id));
+        $checkid = DB::connection('odbc')->select(DB::raw("select * from sabay_items where ID=".$request->id));
                 if(!$checkid){
                         if(!$request->dateadded){$request->dateadded = '0000-00-00';}
-                         DB::connection('odbc')->statement(DB::raw("insert into item (ID,Name,Type,Duration,Price,created_at,updated_at,date_added,group_id) values(".$request->id.",'".$request->name."','".$request->itemtype."',".$request->duration.",".$request->price.",'".$date."','".$date."','".$request->dateadded."',".$request->itemgroup.")"));
+                         DB::connection('odbc')->statement(DB::raw("insert into sabay_items (id,name,type_id,duration,price,created_at,updated_at,date_added,group_id) values(".$request->id.",'".$request->name."','".$request->itemtype."',".$request->duration.",".$request->price.",'".$date."','".$date."','".$request->dateadded."',".$request->itemgroup.")"));
                         }
-                        
                 else{
                     return Redirect::back()->withErrors('ID Already Exist');
                 }
@@ -114,11 +123,20 @@ class ItemController extends Controller
      * @param  int  $id
      * @return Response
      */
-    public function edit($id)
+    public function edit($id,$gametype)
     {
+        $itemtype = SabayItemTypes::lists('name','id');
         $itemgroup = ItemGroup::lists('name','id');
-        $item = Item::findOrFail($id);
-        return view('item.edititem',['item'=>$item,'itemgroup'=>$itemgroup]);
+        if(strcasecmp($gametype, 'ak')==0){
+            $item = Item::findOrFail($id);
+            //dd($item);
+        }
+        elseif(strcasecmp($gametype, 'jx2')==0){
+
+           $items = DB::connection('odbc')->select(DB::raw("select TOP 1 * from sabay_items where id=".$id));
+            foreach($items as $item);
+        }
+        return view('item.edititem',['item'=>$item,'itemgroup'=>$itemgroup,'gametype'=>$gametype,'itemtype'=>$itemtype]);
         
     }
 
@@ -129,13 +147,15 @@ class ItemController extends Controller
      * @param  int  $id
      * @return Response
      */
-    public function update(Request $request, $id)
-    {
-        if(strcasecmp($request->gametypes,'ak') == 0 ){
+    public function update(Request $request, $id,$gametype)
+    {   
+
+        if(strcasecmp($gametype,'ak') == 0 ){
+
                 $item = Item::findOrFail($id);
                 $item->id = $request->id;
                 $item->name = $request->name;
-                $item->type = $request->itemtype;
+                $item->type_id = $request->itemtype;
                 if(strcasecmp($request->itemtype,"periodic") == 0){
                           $item->duration = $request->duration;
                     }
@@ -145,8 +165,11 @@ class ItemController extends Controller
                 $item->group_id = $request->itemgroup;
                 $item->save();
         }
-        elseif(strcasecmp($request->gametypes,'jx2')==0){
-
+        elseif(strcasecmp($gametype,'jx2')==0){
+             $date = date('Y-m-d H:i:s');
+             if(!$request->dateadded){$request->dateadded = '0000-00-00';}
+             DB::connection('odbc')->statement(DB::raw("update sabay_items set name='".$request->name."',type_id='".$request->itemtype."',duration='".$request->duration."',price='".$request->price."',updated_at='".$date."',date_added='".$request->dateadded."'where id=".$id));
+ 
         }
         return Redirect::route('item');
     }
@@ -157,10 +180,15 @@ class ItemController extends Controller
      * @param  int  $id
      * @return Response
      */
-    public function destroy($id)
+    public function destroy($id,$gametype)
     {
+         if(strcasecmp($gametype,'ak') == 0 ){
             $item = Item::findOrFail($id);
             $item->delete();
+            }
+        elseif(strcasecmp($gametype,'jx2') == 0){
+            DB::connection('odbc')->statement(DB::raw("delete from sabay_items where id=".$id));
+        }
             return Redirect::back();
     }
 }
